pyproject.toml
[project]
name = "books-api-cli"
version = "0.1.0"
description = "A CLI tool and API for managing books"
authors = [
    {name = "Christopher", email = "christopherbouda936@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "pyjwt[crypto]>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.6",
]

[project.scripts]
books = "books_api_cli.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-httpx>=0.30.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "B008",  # Allow function calls in argument defaults (required for FastAPI Depends)
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=books_api_cli --cov-report=html --cov-report=term"
